requirements-render.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Login==0.6.3
Werkzeug==2.3.7
SQLAlchemy==2.0.43
python-dotenv==1.0.0
gunicorn==21.2.0
psycopg2-binary==2.9.7  # Nécessaire pour PostgreSQL sur Render
alembic==1.12.0
